Extract FeedInputHeader component
